Develop1 (#3)
* database migration and init_schema are implemented + Makefile for better group development and database name has changed to phonebook

* endpoint naming improved + error message from database is improved

* user secondary port + user secondary actor are implemented + migration up file is enhanced: creating index on username column of user table and defaul value of each table's primary key is auto generated now

* create user dto req and res are implemented + user primary port is implemented

* bcrypt is added for password hashing + user primary port and adaptor are implemented

* password length is fixed + user signup route is implemented

* constants are improved and there is less magic value inside program

* unique constraint of username is handled now

* validating in service layer is improved

* JWT_KEY is added to docker-compose + GenerateToken is added to utils package

* generating jwt for  authorization during user sign up is implemented

* config package variable is localized now and methods are exported functions

* GetUserByUsername in repository is implemented

* user login dtos + GetUserByUsername for service are implemented

* login service is implemented

* Validating request for user login is implemented

* user login controller is implemented

* user login route is implemented

* login controller is implemented

* refresh token generator is implemented

* auh package code is enhanced: duplicated codes are removed for both tokens and auth cookies

* refresh token route is implemented

* checkAccessToken middleware is added and contact resources are behind protected now

* core typed object ids are converted to uuid

* get user id by username is added to secondary adapter contact repository db

* CreateContact method is refactored

* GetAllContacts Method of each user is refactored

* GetContactEmails and GetContactNumbers are refactored

* contact second adapters are totally fixed

* service layer of contacts is fixed

* aat this point, program is running fluently but we have invalid ids issue in url parameter

* id issue is fixed by converting uuid.UUID of app to string and left the database type

* string to uuid convertor function is removed

* EmotyString constant is added to utils package + validation error function is removed

* Access token and refresh token strings literals are constant values now

* log out logic is implemented

* user_role column is added to database schema users table

* user role is added to user repository db port and adapter

* user sevice - sign up new user is updated and compatible with user role

* validating logic of requestBody for createing new user is enhanced + SetUserRole method is improved

* magic value 8 is now behind MinimumPasswordLength constant at utils package

* validating logic of request body for creating new user is improved

+++ src/core/contactRepositoryDb.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"github.com/daniial79/Phone-Book/src/errs"
 	"github.com/daniial79/Phone-Book/src/logger"
-	"strconv"
+	"github.com/daniial79/Phone-Book/src/utils"
 )
 
 // ContactRepositoryDb Secondary actor
@@ -17,7 +17,35 @@ func NewContactRepositoryDb(client *sql.DB) ContactRepositoryDb {
 	return ContactRepositoryDb{client: client}
 }
 
-func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError) {
+// GetContactOwnerByUsername Check for user existence and user id eventually using username
+func (r ContactRepositoryDb) GetContactOwnerByUsername(username string) (string, *errs.AppError) {
+	selectQuery := `SELECT id FROM users WHERE username = $1`
+
+	var userId string
+	row := r.client.QueryRow(selectQuery, username)
+	err := row.Scan(&userId)
+
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return utils.EmptyString, errs.NewNotFoundErr(errs.UserNotFoundErr)
+		}
+
+		return utils.EmptyString, errs.NewUnexpectedErr(errs.InternalErr)
+	}
+
+	return userId, nil
+}
+
+// CreateContact Initiate new contact if owner of contact exists
+func (r ContactRepositoryDb) CreateContact(username string, c *Contact) (*Contact, *errs.AppError) {
+
+	// find contact owner
+	ownerId, appErr := r.GetContactOwnerByUsername(username)
+	if appErr != nil {
+		return nil, appErr
+	}
+	c.OwnerId = ownerId
+
 	insertedNumbers := make([]Number, 0)
 	insertedEmails := make([]Email, 0)
 	tx, err := r.client.Begin()
@@ -25,25 +53,25 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 	//inserting new record to contact tables
 	if err != nil {
 		logger.Error("Error while starting transaction in order to create new contact: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
-	contactInsertSql := `INSERT INTO contacts(first_name, last_name) VALUES($1, $2) RETURNING id`
-	cRow := tx.QueryRow(contactInsertSql, c.FirstName, c.LastName)
+	contactInsertSql := `INSERT INTO contacts(first_name, last_name, owner_id) VALUES($1, $2, $3) RETURNING id`
+	cRow := tx.QueryRow(contactInsertSql, c.FirstName, c.LastName, c.OwnerId)
 
-	var cIntegerId int
-	err = cRow.Scan(&cIntegerId)
+	var insertedContactId string
+	err = cRow.Scan(&insertedContactId)
 	if err != nil {
 		txErr := tx.Rollback()
 		if txErr != nil {
 			logger.Error("Error while rollback the crate contact transaction: " + txErr.Error())
 		}
 		logger.Error("Error while fetching last inserted id from contact tables: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
-	cStringId := strconv.Itoa(cIntegerId)
-	c.Id = cStringId
+	// This will be used for creating contact assets
+	c.Id = insertedContactId
 
 	//inserting new record(s) into numbers
 	insertNumbersSql := `INSERT INTO numbers(contact_id, number, label) VALUES ($1, $2, $3) RETURNING id`
@@ -54,8 +82,8 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 			number.Label,
 		)
 
-		var nIntegerId int
-		err = nRow.Scan(&nIntegerId)
+		var insertedNumberId string
+		err = nRow.Scan(&insertedNumberId)
 
 		if err != nil {
 			txErr := tx.Rollback()
@@ -63,11 +91,10 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 				logger.Error("Error while rollback the crate contact transaction: " + txErr.Error())
 			}
 			logger.Error("Error while fetching last inserted id from numbers tables: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected database error")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 
-		nStringId := strconv.Itoa(nIntegerId)
-		number.Id = nStringId
+		number.Id = insertedNumberId
 		number.ContactId = c.Id
 		insertedNumbers = append(insertedNumbers, number)
 	}
@@ -77,19 +104,18 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 	for _, email := range c.Emails {
 		eRow := tx.QueryRow(insertEmailSql, c.Id, email.Address)
 
-		var eIntegerId int
-		err = eRow.Scan(&eIntegerId)
+		var insertedEmailId string
+		err = eRow.Scan(&insertedEmailId)
 		if err != nil {
 			txErr := tx.Rollback()
 			if txErr != nil {
 				logger.Error("Error while rollback the crate contact transaction: " + txErr.Error())
 			}
 			logger.Error("Error while fetching last inserted id from emails tables: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected database error")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 
-		eStringId := strconv.Itoa(eIntegerId)
-		email.Id = eStringId
+		email.Id = insertedEmailId
 		email.ContactId = c.Id
 
 		insertedEmails = append(insertedEmails, email)
@@ -98,7 +124,7 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 	txErr := tx.Commit()
 	if txErr != nil {
 		logger.Error("Error while committing the new created contact transaction")
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	c.PhoneNumbers = insertedNumbers
@@ -108,16 +134,16 @@ func (r ContactRepositoryDb) CreateContact(c *Contact) (*Contact, *errs.AppError
 }
 
 func (r ContactRepositoryDb) CheckContactExistenceById(cId string) *errs.AppError {
-	var contactId int
+	var contactId string
 	checkContactSql := `SELECT id FROM contacts WHERE id =  $1`
 	row := r.client.QueryRow(checkContactSql, cId)
 	err := row.Scan(&contactId)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return errs.NewNotFoundErr("contact with such id is not found")
+			return errs.NewNotFoundErr(errs.ContactNotFoundErr)
 		}
 		logger.Error("Error while retrieving contact id for existence check (number repo): " + err.Error())
-		return errs.NewUnexpectedErr("Unexpected error happened")
+		return errs.NewUnexpectedErr(errs.InternalErr)
 	}
 	return nil
 }
@@ -136,16 +162,16 @@ func (r ContactRepositoryDb) AddNewNumber(n []Number) ([]Number, *errs.AppError)
 	insertSql := `INSERT INTO numbers(contact_id, number, label) VALUES($1, $2, $3) RETURNING id`
 	for i, number := range n {
 
-		var integerId int
+		var insertedId string
 		row := r.client.QueryRow(insertSql, number.ContactId, number.PhoneNumber, number.Label)
-		err := row.Scan(&integerId)
+		err := row.Scan(&insertedId)
 		if err != nil {
 			logger.Error("Error while retrieving id for last inserted number into existing contact: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected error happened")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 
 		}
-		lastInsertedId := strconv.Itoa(integerId)
-		number.Id = lastInsertedId
+
+		number.Id = insertedId
 
 		result[i] = number
 
@@ -165,16 +191,15 @@ func (r ContactRepositoryDb) AddNewEmails(e []Email) ([]Email, *errs.AppError) {
 	result := make([]Email, len(e))
 	insertSql := `INSERT INTO emails(contact_id, address) VALUES($1, $2) RETURNING id`
 	for i, email := range e {
-		var integerId int
+		var insertedRecordId string
 		row := r.client.QueryRow(insertSql, email.ContactId, email.Address)
-		err := row.Scan(&integerId)
+		err := row.Scan(&insertedRecordId)
 		if err != nil {
 			logger.Error("Error while retrieving id for last inserted email into existing contact:" + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected error happened")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 
-		lastInsertedId := strconv.Itoa(integerId)
-		email.Id = lastInsertedId
+		email.Id = insertedRecordId
 
 		result[i] = email
 	}
@@ -182,14 +207,19 @@ func (r ContactRepositoryDb) AddNewEmails(e []Email) ([]Email, *errs.AppError) {
 	return result, nil
 }
 
-func (r ContactRepositoryDb) GetAllContacts() ([]Contact, *errs.AppError) {
-	contacts := make([]Contact, 0)
-	selectContactSql := `SELECT * FROM contacts`
+func (r ContactRepositoryDb) GetAllContacts(username string) ([]Contact, *errs.AppError) {
+	ownerId, appErr := r.GetContactOwnerByUsername(username)
+	if appErr != nil {
+		return nil, errs.NewNotFoundErr(errs.UserNotFoundErr)
+	}
 
-	rows, err := r.client.Query(selectContactSql)
+	contacts := make([]Contact, 0)
+	selectContactSql := `SELECT id, first_name, last_name FROM contacts WHERE owner_id = $1`
+
+	rows, err := r.client.Query(selectContactSql, ownerId)
 	if err != nil {
 		logger.Error("Error while querying contacts table: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	for rows.Next() {
@@ -197,7 +227,7 @@ func (r ContactRepositoryDb) GetAllContacts() ([]Contact, *errs.AppError) {
 		err = rows.Scan(&c.Id, &c.FirstName, &c.LastName)
 		if err != nil {
 			logger.Error("Error while scanning retrieved records from contacts table: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected database error")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 		contacts = append(contacts, c)
 	}
@@ -216,7 +246,7 @@ func (r ContactRepositoryDb) GetContactNumbers(cId string) ([]Number, *errs.AppE
 			return nil, errs.NewNotFoundErr("no number associated with this contact id")
 		}
 		logger.Error("Error while selecting numbers associated with specific contact: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	for rows.Next() {
@@ -224,7 +254,7 @@ func (r ContactRepositoryDb) GetContactNumbers(cId string) ([]Number, *errs.AppE
 		err = rows.Scan(&n.Id, &n.ContactId, &n.PhoneNumber, &n.Label)
 		if err != nil {
 			logger.Error("Error while scanning retrieved set of numbers: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected database error")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 		numbers = append(numbers, n)
 	}
@@ -243,7 +273,7 @@ func (r ContactRepositoryDb) GetContactEmails(cId string) ([]Email, *errs.AppErr
 			return nil, errs.NewNotFoundErr("no email associated with this contact id")
 		}
 		logger.Error("Error while selecting emails associated with specific contact: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	for rows.Next() {
@@ -251,7 +281,7 @@ func (r ContactRepositoryDb) GetContactEmails(cId string) ([]Email, *errs.AppErr
 		err = rows.Scan(&e.Id, &e.ContactId, &e.Address)
 		if err != nil {
 			logger.Error("Error while scanning retrieved set of emails: " + err.Error())
-			return nil, errs.NewUnexpectedErr("Unexpected database error")
+			return nil, errs.NewUnexpectedErr(errs.InternalErr)
 		}
 		emails = append(emails, e)
 	}
@@ -268,10 +298,10 @@ func (r ContactRepositoryDb) DeleteContactEmail(cId, eId string) *errs.AppError 
 	err := row.Scan(&deletedEmailId)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return errs.NewNotFoundErr("email not found")
+			return errs.NewNotFoundErr(errs.EmailNotFoundErr)
 		}
 		logger.Error("Error while removing a record from emails table: " + err.Error())
-		return errs.NewUnexpectedErr("Unexpected database error")
+		return errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	return nil
@@ -286,19 +316,20 @@ func (r ContactRepositoryDb) DeleteContactPhoneNumber(cId, nId string) *errs.App
 	err := row.Scan(&deletedNumberId)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return errs.NewNotFoundErr("phone number not found")
+			return errs.NewNotFoundErr(errs.NumberNotFoundErr)
 		}
-		return errs.NewUnexpectedErr("Unexpected database error")
+		return errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	return nil
 }
 
 func (r ContactRepositoryDb) DeleteContact(cId string) *errs.AppError {
+
 	tx, err := r.client.Begin()
 	if err != nil {
 		logger.Error("Error while starting new transaction for cascading deletion of contact: " + err.Error())
-		return errs.NewUnexpectedErr("Unexpected database error")
+		return errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	pnDeleteQuery := `DELETE FROM numbers WHERE contact_id = $1`
@@ -309,7 +340,7 @@ func (r ContactRepositoryDb) DeleteContact(cId string) *errs.AppError {
 		txErr := tx.Rollback()
 		if txErr != nil {
 			logger.Error("Error while rollback from cascading deletion on numbers table:" + txErr.Error())
-			return errs.NewUnexpectedErr("Unexpected database error")
+			return errs.NewUnexpectedErr(errs.InternalErr)
 		}
 	}
 
@@ -321,11 +352,11 @@ func (r ContactRepositoryDb) DeleteContact(cId string) *errs.AppError {
 		txErr := tx.Rollback()
 		if txErr != nil {
 			logger.Error("Error while rollback from cascading deletion on emails table:" + txErr.Error())
-			return errs.NewUnexpectedErr("Unexpected database error")
+			return errs.NewUnexpectedErr(errs.InternalErr)
 		}
 	}
 
-	cDeleteQuery := `DELETE FROM contacts WHERE id = $1 RETURNING id`
+	cDeleteQuery := `DELETE FROM contacts WHERE id = $1 AND owner_id = $2 RETURNING id`
 	row := tx.QueryRow(cDeleteQuery, cId)
 	var deletedContactId int
 
@@ -336,7 +367,7 @@ func (r ContactRepositoryDb) DeleteContact(cId string) *errs.AppError {
 
 		if txErr := tx.Rollback(); txErr != nil {
 			logger.Error("Error while rollback from deleting record from contacts table:" + txErr.Error())
-			return errs.NewUnexpectedErr("Unexpected database error")
+			return errs.NewUnexpectedErr(errs.InternalErr)
 		}
 	}
 
@@ -358,25 +389,25 @@ func (r ContactRepositoryDb) UpdateContactPhoneNumber(newNumber Number) (*Number
 	}
 
 	var row *sql.Row
-	if newPhoneNumber != "" && newLabel != "" {
+	if newPhoneNumber != utils.EmptyString && newLabel != utils.EmptyString {
 		updateQuery := `UPDATE numbers SET number = $1, label = $2 WHERE id = $3 AND contact_id = $4 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newPhoneNumber, newLabel, numberId, contactId)
-	} else if newPhoneNumber == "" && newLabel != "" {
+	} else if newPhoneNumber == utils.EmptyString && newLabel != utils.EmptyString {
 		updateQuery := `UPDATE numbers SET label = $1 WHERE id = $2 AND contact_id = $3 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newLabel, numberId, contactId)
-	} else if newPhoneNumber != "" && newLabel == "" {
+	} else if newPhoneNumber != utils.EmptyString && newLabel == utils.EmptyString {
 		updateQuery := `UPDATE numbers SET number = $1 WHERE id = $2 AND contact_id = $3 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newPhoneNumber, numberId, contactId)
 	}
 
-	var retrievedId int
+	var retrievedId string
 	err := row.Scan(&retrievedId)
 	if err != nil {
 		if errors.Is(sql.ErrNoRows, err) {
-			return nil, errs.NewNotFoundErr("phone number with this id not found")
+			return nil, errs.NewNotFoundErr(errs.NumberNotFoundErr)
 		}
 		logger.Error("Error while updating record in numbers table: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	return &newNumber, nil
@@ -393,14 +424,14 @@ func (r ContactRepositoryDb) UpdateContactEmail(newEmail Email) (*Email, *errs.A
 
 	updateQuery := `UPDATE emails SET address = $1 WHERE id = $2 AND contact_id = $3 RETURNING id`
 	row := r.client.QueryRow(updateQuery, newAddress, emailId, contactId)
-	var retrievedId int
+	var retrievedId string
 	err := row.Scan(&retrievedId)
 	if err != nil {
 		if errors.Is(sql.ErrNoRows, err) {
 			return nil, errs.NewNotFoundErr("email with this id not found")
 		}
 		logger.Error("Error while updating record in emails table: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	return &newEmail, nil
@@ -416,23 +447,22 @@ func (r ContactRepositoryDb) UpdateContact(newContact Contact) (*Contact, *errs.
 	}
 
 	var row *sql.Row
-	if newFirstname != "" && newLastname != "" {
+	if newFirstname != utils.EmptyString && newLastname != utils.EmptyString {
 		updateQuery := `UPDATE contacts SET first_name = $1, last_name = $2 WHERE id = $3 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newFirstname, newLastname, contactId)
-	} else if newFirstname != "" && newLastname == "" {
+	} else if newFirstname != utils.EmptyString && newLastname == utils.EmptyString {
 		updateQuery := `UPDATE contacts SET first_name = $1 WHERE id = $2 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newFirstname, contactId)
-	} else if newFirstname == "" && newLastname != "" {
+	} else if newFirstname == utils.EmptyString && newLastname != utils.EmptyString {
 		updateQuery := `UPDATE contacts SET last_name = $1 WHERE id = $2 RETURNING id`
 		row = r.client.QueryRow(updateQuery, newLastname, contactId)
 	}
 
-	var retrievedId int
+	var retrievedId string
 	err := row.Scan(&retrievedId)
 	if err != nil {
-		//at this stage, we would not face empty record as result
 		logger.Error("Error while updating record in contact tables: " + err.Error())
-		return nil, errs.NewUnexpectedErr("Unexpected database error")
+		return nil, errs.NewUnexpectedErr(errs.InternalErr)
 	}
 
 	return &newContact, nil
